+++ CHANGES.rst
@@ -4,7 +4,8 @@ Werkzeug Changelog
 Version 0.14
 ------------
 
-Unreleased
+- HTTP execptions are now automatically caught by
+  ``Request.application``.
 
 
 Version 0.13